+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - API/CLI
   - Support renaming namespaces as well with `rename` feature.
   - Use relative paths instead of absolute paths on diff messages.
+  - Add `analyze-project!` to analyze whole project only and cache analysis, useful for REPL usage for example. 
 
 ## 2021.09.13-22.25.35